docker/flow-ai/requirements.txt
torch==2.2.2
diffusers==0.27.2
transformers==4.40.1
accelerate==0.29.3
safetensors==0.4.3
fastapi==0.111.0
uvicorn[standard]==0.29.0
Pillow==10.3.0
numpy<2
huggingface-hub==0.23.5
optimum[onnxruntime]==1.19.2
onnxruntime==1.18.1
